backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
python-jose[cryptography]
python-multipart
httpx
pytest